+++ zython/__init__.py
@@ -3,7 +3,7 @@ from zython.var_par.var import var
 from zython.var_par.par import par
 from zython.var_par.array import Array
 from zython.operations.functions_and_predicates import exists, forall, sum, alldifferent, circuit, count, min, max,\
-    allequal, ndistinct
+    allequal, ndistinct, increasing, decreasing
 from zython.model import Model
 from zython.result import as_original