+++ zython/operations/constraint.py
@@ -1,6 +1,6 @@
-import zython
-from typing import Union, Optional, Callable, Sequence
+from typing import Union, Optional, Callable
 
+import zython
 from zython.operations._op_codes import _Op_code
 
 
@@ -26,20 +26,14 @@ class Constraint:
     def __and__(self, other):
         return Constraint(_Op_code.and_, self, other)
 
-    @staticmethod
-    def exists(seq: Union["zython.var_par.types._range",
-                          "zython.var_par.types.orig_range",
-                          "zython.var_par.array.ArrayMixin",
-                          Sequence["zython.var"]],
-               iter_var: Optional["zython.var"] = None,
-               func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
-        return Constraint(_Op_code.exists, seq, iter_var, func)
 
-    @staticmethod
-    def forall(seq: Union["zython.var_par.types._range",
-                          "zython.var_par.types.orig_range",
-                          "zython.var_par.array.ArrayMixin",
-                          Sequence["zython.var"]],
-               iter_var: Optional["zython.var"] = None,
-               func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
-        return Constraint(_Op_code.forall, seq, iter_var, func)
+def _exists(seq: "zython.var_par.types.ZnSequence",
+            iter_var: Optional["zython.var"] = None,
+            func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
+    return Constraint(_Op_code.exists, seq, iter_var, func)
+
+
+def _forall(seq: "zython.var_par.types.ZnSequence",
+            iter_var: Optional["zython.var"] = None,
+            func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
+    return Constraint(_Op_code.forall, seq, iter_var, func)