+++ changelog.md
@@ -1,3 +1,11 @@
+### 0.1.2
+#### Fixed
+- some method of Operation and Constraint classes which were
+  accessible by and visible for user are now hidden
+#### Added
+- ``increasing`` and ``decreasing`` constraints
+  
+  
 ### 0.1.1
 #### Added
 - ``allequal`` constraint